move snackbar to features
